refactor: move wallet state screens to the `entities` folder
