+++ packages/dialect-react-ui/entities/wallet-states/SignMessageInfo.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
-import { Centered, Loader } from '../../common';
-import { H3, P } from '../../common/preflighted';
-import { useTheme } from '../../common/providers/DialectThemeProvider';
+import { Centered, Loader } from '../../components/common';
+import { H3, P } from '../../components/common/preflighted';
+import { useTheme } from '../../components/common/providers/DialectThemeProvider';
 
 const SignMessageInfo = () => {
   const { textStyles } = useTheme();
@@ -12,7 +12,7 @@ const SignMessageInfo = () => {
       >
         <Loader className="dt-mr-2" /> Waiting for your wallet
       </H3>
-      <P className="dt-text-center dt-max-w-sm dt-min-w-[24rem] dt-opacity-50">
+      <P className="dt-text-center dt-max-w-sm dt-opacity-50">
         To continue please prove you own a wallet by approving signing request.
         It is free and does not involve the network.
       </P>
